syllabus/requirements.txt
pantable==0.13.3
python-dateutil==2.8.1
